pyproject.toml
[project]
name = "solar-registry"
version = "0.1.0"
description = "Generate and merge meta data for testsolar test tools"
authors = [
    {name = "asiazhang", email = "asiazhang2002@gmail.com"},
]
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.7.1",
    "typer>=0.12.3",
    "pyyaml>=6.0.1",
    "retry>=0.9.2",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.2.0",
]
